Make qRound explicitly choose the double overload

+++ sheets/Value.cpp
@@ -499,7 +499,7 @@ QDateTime Value::asDateTime(const CalculationSettings* settings) const
     QDateTime datetime(settings->referenceDate(), QTime(), Qt::UTC);
 
     const int days = asInteger();
-    const int msecs = qRound((numToDouble(asFloat() - double(days))) * 86400000.0);      // 24*60*60*1000
+    const int msecs = qRound((qreal)(numToDouble(asFloat() - double(days))) * 86400000.0);      // 24*60*60*1000
     datetime = datetime.addDays(days);
     datetime = datetime.addMSecs(msecs);
 
@@ -524,7 +524,7 @@ QTime Value::asTime(const CalculationSettings* settings) const
     QTime dt;
 
     const int days = asInteger();
-    const int msecs = qRound(numToDouble(asFloat() - double(days)) * 86400000.0);      // 24*60*60*1000
+    const int msecs = qRound((qreal)numToDouble(asFloat() - double(days)) * 86400000.0);      // 24*60*60*1000
     dt = dt.addMSecs(msecs);
 
     return dt;
